Toggle revision.
