+++ webview-ui/src/components/history/HistoryPreview.tsx
@@ -1,52 +1,23 @@
-import { memo, useState, useCallback } from "react"
+import { memo } from "react"
 
 import { vscode } from "@/utils/vscode"
 import { formatLargeNumber, formatDate } from "@/utils/format"
 import { Button } from "@/components/ui"
 
 import { useAppTranslation } from "@src/i18n/TranslationContext"
-import { useExtensionState } from "@src/context/ExtensionStateContext"
 import { CopyButton } from "./CopyButton"
 import { useTaskSearch } from "./useTaskSearch"
 
 import { Trans } from "react-i18next"
 import { Coins } from "lucide-react"
 
-type HistoryPreviewProps = {
-	showHistoryView: () => void
-}
-const HistoryPreview = ({ showHistoryView }: HistoryPreviewProps) => {
+const HistoryPreview = () => {
 	const { tasks, showAllWorkspaces } = useTaskSearch()
-	const { historyPreviewCollapsed } = useExtensionState() // Will add this state later
 	const { t } = useAppTranslation()
-	// Initialize expanded state based on the persisted setting (default to expanded if undefined)
-	const [isExpanded, setIsExpanded] = useState(
-		historyPreviewCollapsed === undefined ? true : !historyPreviewCollapsed,
-	)
-
-	const toggleExpanded = useCallback(() => {
-		const newState = !isExpanded
-		setIsExpanded(newState)
-		// Send message to extension to persist the new collapsed state
-		vscode.postMessage({ type: "setHistoryPreviewCollapsed", bool: !newState })
-	}, [isExpanded])
 
 	return (
 		<>
 			<div className="flex flex-col gap-3 shrink-0 mx-5">
-				{tasks.length !== 0 && (
-					<div className="flex items-center justify-between text-vscode-descriptionForeground  w-full mx-auto max-w-[600px]">
-						<div className="flex items-center gap-1 cursor-pointer" onClick={toggleExpanded}>
-							<span className={`codicon codicon-chevron-${isExpanded ? "down" : "right"} scale-90`} />
-							<span className="font-medium text-xs uppercase">{t("history:recentTasks")}</span>
-						</div>
-						{/* Keep the history button, but maybe it should just show the full view? Or remove it if header is clicked? Let's keep it for now. */}
-						<Button variant="ghost" size="sm" onClick={() => showHistoryView()} className="uppercase">
-							<span className="codicon codicon-history size-[1rem]" />
-						</Button>
-					</div>
-				)}
-
 				{tasks.length === 0 && (
 					<>
 						<p className="opacity-50 p-2 text-center my-0 mx-auto max-w-80">
@@ -65,14 +36,21 @@ const HistoryPreview = ({ showHistoryView }: HistoryPreviewProps) => {
 							/>
 						</p>
 
-						<Button size="sm" variant="secondary" onClick={() => showHistoryView()} className="mx-auto">
+						{/* The button to show history view will be moved to ChatView */}
+						<Button
+							size="sm"
+							variant="secondary"
+							onClick={() => {
+								/* TODO: Implement or remove */
+							}}
+							className="mx-auto">
 							<span className="codicon codicon-history size-[1rem]" />
 							{t("history:viewAll")}
 						</Button>
 					</>
 				)}
 
-				{tasks.length !== 0 && isExpanded && (
+				{tasks.length !== 0 && (
 					<>
 						{tasks.slice(0, 3).map((item) => (
 							<div
@@ -116,10 +94,12 @@ const HistoryPreview = ({ showHistoryView }: HistoryPreviewProps) => {
 								</div>
 							</div>
 						))}
-						{/* Add a "View All" link below the preview list when expanded */}
+						{/* The link to show history view will be moved to ChatView */}
 						<div
 							className="text-center text-xs text-vscode-descriptionForeground cursor-pointer hover:text-vscode-foreground mt-1"
-							onClick={() => showHistoryView()}>
+							onClick={() => {
+								/* TODO: Implement or remove */
+							}}>
 							{t("history:viewAll")} ({tasks.length})
 						</div>
 					</>
